requirements.txt
Flask==3.0.3
pdfminer.six==20221105
python-docx==1.1.0
spacy==3.7.4
scikit-learn==1.4.2
pandas==2.2.2
gunicorn==21.2.0
